DeleteMessage returns wrong error, 0-byte Auth Header value driven by a const
The DeleteMessage queue method also returns a 404 instead of a 404 (only tested against local emulator in this case). 

The authe header was upset at an empty content length (which used to work? prior API or emulator version maybe?) so I added a constant to switch it between "0", which a newer emulator likes, and "" if it is needed for other environments.